Issue 16 move tabs to separate controllers
