+++ backupper-gui/src/main/java/pl/gda/pg/eti/autyzm/Controller/BaseController.java
@@ -1,0 +1,38 @@
+package pl.gda.pg.eti.autyzm.Controller;
+
+import pl.gda.pg.eti.autyzm.Model.Info;
+import pl.gda.pg.eti.autyzm.StringConfig;
+import pl.gda.pg.eti.autyzm.backupper.core.AdbProxy;
+import se.vidstige.jadb.JadbDevice;
+import se.vidstige.jadb.JadbException;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Created by malgorzatas on 30/08/16.
+ */
+public abstract class BaseController {
+
+    protected void initAdbConnection() {
+        try {
+            AdbProxy.initAdbConnection();
+        } catch (IOException e) {
+            e.printStackTrace();
+            Info.showAlert(StringConfig.FAILED_TO_INIT_ADB_CONNECTION_TITLE, StringConfig.FAILED_TO_INIT_ADB_CONNECTION_BODY,
+                    null, Info.TYPE.ERROR);
+        }
+    }
+
+    protected List<JadbDevice> getConnectedDevices() {
+        try {
+            return AdbProxy.getConnectedDevices();
+        } catch (IOException | JadbException e) {
+            Info.showAlert(StringConfig.LOOKING_FOR_DEVICES_ERROR_TITLE, StringConfig.LOOKING_FOR_DEVICES_ERROR_BODY,
+                    null, Info.TYPE.ERROR);
+            e.printStackTrace();
+            return new ArrayList<>();
+        }
+    }
+}
