+++ backupper-gui/src/main/java/pl/gda/pg/eti/autyzm/Controller/MainController.java
@@ -1,0 +1,15 @@
+package pl.gda.pg.eti.autyzm.Controller;
+
+import javafx.fxml.FXML;
+
+public class MainController {
+
+    @FXML private BackupController backupController;
+    @FXML private RestoreController restoreController;
+
+    @FXML
+    public void initialize() {
+    }
+
+
+}
