+++ backupper-gui/src/main/resources/backup.fxml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TableColumn?>
+<?import pl.gda.pg.eti.autyzm.StringConfig?>
+
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.image.Image?>
+<BorderPane xmlns:fx="http://javafx.com/fxml/1"
+            fx:controller="pl.gda.pg.eti.autyzm.Controller.BackupController">
+<center>
+<VBox styleClass="copyTab">
+    <VBox styleClass="copyNameInputArea">
+        <Label>
+            <StringConfig fx:constant="COPY_INPUT_NAME_LABEL"/>
+        </Label>
+        <TextField fx:id="nameInput"/>
+    </VBox>
+    <VBox>
+        <Button onAction="#refreshDeviceList">
+            <graphic>
+                <ImageView fitHeight="20.0" fitWidth="20.0">
+                    <image>
+                        <Image url="images/refresh.png" />
+                    </image>
+                </ImageView>
+            </graphic>
+        </Button>
+        <HBox styleClass="chooseDeviceArea">
+            <TableView fx:id="deviceTableView">
+                <columns>
+                    <TableColumn fx:id="device">
+                        <text>
+                            <StringConfig fx:constant="DEVICE_COLUMN"/>
+                        </text>
+                    </TableColumn>
+                    <TableColumn fx:id="chooseDevice" sortable="false">
+                    </TableColumn>
+                </columns>
+            </TableView>
+        </HBox>
+    </VBox>
+</VBox>
+</center>
+</BorderPane>
